requirements.txt
asgiref==3.7.2
Brotli==1.1.0
Django==3.2
djangorestframework==3.14.0
gunicorn==21.2.0
mssql-django==1.3
packaging==23.2
Pillow==10.1.0
pyodbc==5.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
whitenoise==6.6.0
